add CI for pull requests
Note that this uses the `pull_request` event so it runs the code in the PR branch and does not have access to repo secrets (I don't think there are any relevant ones here though). We need to be extra careful here with external PRs because we run on self-hosted runners.